#11512: Rebase refactoring branch
